backend/requirements.txt
boto3==1.35.15
Flask==2.0.3
Werkzeug==2.0.3
Flask_Cors==5.0.0
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
